+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 Next Release
 -
-	
+* If Statements and Comparison Operators
 
 Version 1.0.0
 -